+++ src/main/kotlin/icu/windea/pls/lang/resolve/complexExpression/nodes/ParadoxStaticCommandFieldNode.kt
@@ -15,6 +15,7 @@ import icu.windea.pls.lang.isParameterized
 import icu.windea.pls.lang.psi.ParadoxExpressionElement
 import icu.windea.pls.lang.references.CwtConfigBasedPsiReference
 import icu.windea.pls.lang.util.ParadoxExpressionManager
+import icu.windea.pls.model.constraints.ParadoxReferenceConstraint
 
 class ParadoxStaticCommandFieldNode(
     override val text: String,