+++ src/main/kotlin/icu/windea/pls/lang/resolve/complexExpression/nodes/ParadoxStaticCommandScopeNode.kt
@@ -16,6 +16,7 @@ import icu.windea.pls.lang.isParameterized
 import icu.windea.pls.lang.psi.ParadoxExpressionElement
 import icu.windea.pls.lang.references.CwtConfigBasedPsiReference
 import icu.windea.pls.lang.util.ParadoxExpressionManager
+import icu.windea.pls.model.constraints.ParadoxReferenceConstraint
 
 class ParadoxStaticCommandScopeNode(
     override val text: String,