+++ src/main/kotlin/icu/windea/pls/model/constraints/ParadoxReferenceConstraint.kt
@@ -2,6 +2,7 @@ package icu.windea.pls.model.constraints
 
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
+import icu.windea.pls.config.CwtDataType
 import icu.windea.pls.config.CwtDataTypeSets
 import icu.windea.pls.config.CwtDataTypes
 import icu.windea.pls.csv.psi.ParadoxCsvColumn
@@ -12,8 +13,8 @@ import icu.windea.pls.lang.psi.isComplexExpression
 import icu.windea.pls.lang.psi.isDatabaseObjectExpression
 import icu.windea.pls.lang.psi.isResolvableLiteralExpression
 import icu.windea.pls.lang.references.ParadoxComplexEnumValuePsiReference
+import icu.windea.pls.lang.references.ParadoxConstrainedPsiReference
 import icu.windea.pls.lang.references.ParadoxScriptedVariablePsiReference
-import icu.windea.pls.lang.references.csv.ParadoxCsvExpressionPsiReference
 import icu.windea.pls.lang.references.localisation.ParadoxLocalisationConceptPsiReference
 import icu.windea.pls.lang.references.localisation.ParadoxLocalisationIconPsiReference
 import icu.windea.pls.lang.references.localisation.ParadoxLocalisationParameterPsiReference
@@ -24,8 +25,6 @@ import icu.windea.pls.lang.references.script.ParadoxConditionParameterPsiReferen
 import icu.windea.pls.lang.references.script.ParadoxDefinitionInjectionTargetPsiReference
 import icu.windea.pls.lang.references.script.ParadoxEventNamespacePsiReference
 import icu.windea.pls.lang.references.script.ParadoxParameterPsiReference
-import icu.windea.pls.lang.references.script.ParadoxScriptExpressionPsiReference
-import icu.windea.pls.lang.resolve.complexExpression.nodes.*
 import icu.windea.pls.localisation.psi.ParadoxLocalisationConceptCommand
 import icu.windea.pls.localisation.psi.ParadoxLocalisationExpressionElement
 import icu.windea.pls.localisation.psi.ParadoxLocalisationIcon
@@ -57,17 +56,16 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
                 is ParadoxScriptedVariablePsiReference -> true
                 else -> false
             }
         }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            return false // unavailable
+        }
     },
 
-    /**
-     * @see CwtDataTypeSets.DefinitionAware
-     * @see CwtDataTypes.AliasKeysField
-     */
     Definition {
         override fun canResolveReference(element: PsiElement): Boolean {
             return when (element) {
@@ -85,36 +83,23 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
-                is ParadoxScriptExpressionPsiReference -> {
-                    reference.configs.any { config ->
-                        val configExpression = config.configExpression
-                        val dataType = configExpression.type
-                        dataType in CwtDataTypeSets.DefinitionAware || dataType == CwtDataTypes.AliasKeysField
-                    }
-                }
+                is ParadoxConstrainedPsiReference -> reference.canResolveFor(this)
                 is ParadoxLocalisationTextColorPsiReference -> true // <text_color>
                 is ParadoxLocalisationIconPsiReference -> true // <sprite>, etc.
                 is ParadoxLocalisationConceptPsiReference -> true // <game_concept>
                 is ParadoxLocalisationTextIconPsiReference -> true // <text_icon>
                 is ParadoxLocalisationTextFormatPsiReference -> true // <text_format>
-                is ParadoxCsvExpressionPsiReference -> {
-                    val config = reference.columnConfig.valueConfig ?: return false
-                    val configExpression = config.configExpression
-                    val dataType = configExpression.type
-                    dataType == CwtDataTypes.Definition
-                }
                 is ParadoxEventNamespacePsiReference -> true // <event_namespace>
                 is ParadoxDefinitionInjectionTargetPsiReference -> true
                 else -> false
             }
         }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            return dataType in CwtDataTypeSets.DefinitionAware || dataType in CwtDataTypeSets.Expandable
+        }
     },
 
-    /**
-     * @see CwtDataTypeSets.NormalLocalisationAware
-     * @see CwtDataTypes.AliasKeysField
-     */
     Localisation {
         override fun canResolveReference(element: PsiElement): Boolean {
             return when (element) {
@@ -127,24 +112,17 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
-                is ParadoxScriptExpressionPsiReference -> {
-                    reference.configs.any { config ->
-                        val configExpression = config.configExpression
-                        val dataType = configExpression.type
-                        dataType in CwtDataTypeSets.NormalLocalisationAware || dataType == CwtDataTypes.AliasKeysField
-                    }
-                }
+                is ParadoxConstrainedPsiReference -> reference.canResolveFor(this)
                 is ParadoxLocalisationParameterPsiReference -> true
                 else -> false
             }
         }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            return dataType in CwtDataTypeSets.LocalisationAware || dataType in CwtDataTypeSets.Expandable
+        }
     },
 
-    /**
-     * @see CwtDataTypes.EnumValue
-     * @see CwtDataTypes.AliasKeysField
-     */
     ComplexEnumValue {
         override fun canResolveReference(element: PsiElement): Boolean {
             return when (element) {
@@ -156,30 +134,17 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
+                is ParadoxConstrainedPsiReference -> reference.canResolveFor(this)
                 is ParadoxComplexEnumValuePsiReference -> true
-                is ParadoxScriptExpressionPsiReference -> {
-                    reference.configs.any { config ->
-                        val configExpression = config.configExpression
-                        val dataType = configExpression.type
-                        dataType == CwtDataTypes.EnumValue || dataType == CwtDataTypes.AliasKeysField
-                    }
-                }
-                is ParadoxCsvExpressionPsiReference -> {
-                    val config = reference.columnConfig.valueConfig ?: return false
-                    val configExpression = config.configExpression
-                    val dataType = configExpression.type
-                    dataType == CwtDataTypes.EnumValue
-                }
                 else -> false
             }
         }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            return dataType == CwtDataTypes.EnumValue || dataType in CwtDataTypeSets.Expandable
+        }
     },
 
-    /**
-     * @see CwtDataTypeSets.DynamicValue
-     * @see CwtDataTypes.AliasKeysField
-     */
     DynamicValue {
         override fun canResolveReference(element: PsiElement): Boolean {
             return when (element) {
@@ -192,28 +157,16 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
-                is ParadoxScriptExpressionPsiReference -> {
-                    reference.configs.any { config ->
-                        val configExpression = config.configExpression
-                        val dataType = configExpression.type
-                        dataType in CwtDataTypeSets.DynamicValue || dataType == CwtDataTypes.AliasKeysField
-                    }
-                }
-                is ParadoxCsvExpressionPsiReference -> {
-                    val config = reference.columnConfig.valueConfig ?: return false
-                    val configExpression = config.configExpression
-                    val dataType = configExpression.type
-                    dataType in CwtDataTypeSets.DynamicValue
-                }
+                is ParadoxConstrainedPsiReference -> reference.canResolveFor(this)
                 else -> false
             }
         }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            return dataType in CwtDataTypeSets.DynamicValue || dataType in CwtDataTypeSets.Expandable
+        }
     },
 
-    /**
-     * @see CwtDataTypes.Parameter
-     */
     Parameter {
         override fun canResolveReference(element: PsiElement): Boolean {
             return when (element) {
@@ -226,24 +179,18 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
-                is ParadoxScriptExpressionPsiReference -> {
-                    reference.configs.any { config ->
-                        val configExpression = config.configExpression
-                        val dataType = configExpression.type
-                        dataType == CwtDataTypes.Parameter
-                    }
-                }
+                is ParadoxConstrainedPsiReference -> reference.canResolveFor(this)
                 is ParadoxParameterPsiReference -> true
                 is ParadoxConditionParameterPsiReference -> true
                 else -> false
             }
         }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            return dataType == CwtDataTypes.Parameter || dataType in CwtDataTypeSets.Expandable
+        }
     },
 
-    /**
-     * @see CwtDataTypes.LocalisationParameter
-     */
     LocalisationParameter {
         override fun canResolveReference(element: PsiElement): Boolean {
             return when (element) {
@@ -255,24 +202,17 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
-                is ParadoxScriptExpressionPsiReference -> {
-                    reference.configs.any { config ->
-                        val configExpression = config.configExpression
-                        val dataType = configExpression.type
-                        dataType == CwtDataTypes.LocalisationParameter
-                    }
-                }
+                is ParadoxConstrainedPsiReference -> reference.canResolveFor(this)
                 is ParadoxLocalisationParameterPsiReference -> true
                 else -> false
             }
         }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            return dataType == CwtDataTypes.LocalisationParameter || dataType in CwtDataTypeSets.Expandable
+        }
     },
 
-    /**
-     * @see CwtDataTypeSets.LocalisationReference
-     * @see CwtDataTypes.AliasKeysField
-     */
     LocalisationReference {
         override fun canResolveReference(element: PsiElement): Boolean {
             return Localisation.canResolveReference(element)
@@ -280,17 +220,14 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
-                is ParadoxScriptExpressionPsiReference -> {
-                    // also for synced_localisation
-                    reference.configs.any { config ->
-                        val configExpression = config.configExpression
-                        val dataType = configExpression.type
-                        dataType in CwtDataTypeSets.LocalisationReference || dataType == CwtDataTypes.AliasKeysField
-                    }
-                }
+                is ParadoxConstrainedPsiReference -> reference.canResolveFor(this)
                 else -> Localisation.canResolve(reference)
             }
+        }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            // also for synced_localisation
+            return dataType in CwtDataTypeSets.LocalisationReference || dataType in CwtDataTypeSets.Expandable
         }
     },
 
@@ -301,14 +238,21 @@ enum class ParadoxReferenceConstraint {
 
         override fun canResolve(reference: PsiReference): Boolean {
             return when (reference) {
-                is ParadoxIdentifierNode.Reference -> reference.canResolveFor(this)
+                is ParadoxConstrainedPsiReference -> reference.canResolveFor(this)
                 else -> DynamicValue.canResolve(reference)
             }
         }
+
+        override fun test(dataType: CwtDataType): Boolean {
+            return dataType in CwtDataTypeSets.DynamicValue || dataType in CwtDataTypeSets.Expandable
+        }
     },
+
     ;
 
     open fun canResolveReference(element: PsiElement): Boolean = true
 
     open fun canResolve(reference: PsiReference): Boolean = true
+
+    open fun test(dataType: CwtDataType): Boolean = true
 }